refactor(relay): split engine.go god-object into strategy files (#229)
engine.go was 1532 LOC mixing 5 connection strategies + audio config
parsing. Split along strategy boundaries:

- engine.go (296 LOC, was 1532): PushTarget struct, constructor, Set*
  config setters, Status, Run reconnect loop, connectAndStream
  dispatcher, setStatus
- connect_ffmpeg.go (150 LOC): connectViaFFmpeg + probeSourceVideoFPS
  (FFmpeg subprocess fallback for strict RTMP receivers)
- connect_rtmp.go (264 LOC): connectRTMP (native Go RTMP writer with
  AAC/G.711/silent-audio handling)
- connect_rtsp.go (244 LOC): connectRTSP (gortsplib recording mode)
- connect_transcode.go (570 LOC): connectRTMPWithTranscode +
  connectRTSPWithTranscode + waitForTranscoderParams + relayPresetToLT
  (H.265→H.264 transcode-wrapped variants)
- audio_config.go (55 LOC): parseASC + parseG711Config + durationFromPTS
  + errPermanent sentinel

No behavioral change: all methods moved verbatim with original comments,
all signatures preserved. The connectAndStream dispatcher fans out to
the strategy files unchanged.

Verified: go build ./... passes; go test -race ./internal/relay/... passes
(22s); coverage 25.5% maintained (split is mechanical; mock-RTMP/RTSP
tests to reach the 60% target are a separate follow-up per the issue);
go vet + gofmt clean.

Closes #229
